modules-load: implement parallel module loading
Load modules in parallel using a pool of worker threads. The number of
threads is equal to the number of CPUs, with a maximum of 8 (to avoid
too many probes running in parallel).

The number of thread can optionally be specified manually using the
SYSTEMD_MODULES_LOAD_NUM_CONCURRENT environment variable; in this case,
no limit is enforced.

Signed-off-by: Francesco Valla <francesco@valla.it>